backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pydantic[email]==2.9.2
email-validator==2.2.0
aiofiles==24.1.0
psycopg2-binary==2.9.10
